Put classes back in one file
