+++ safe_power_monitor.py
@@ -1,7 +1,3 @@
-import ptvsd
-
-ptvsd.enable_attach('secret')
-
 #!/usr/bin/env python2.7
 # date: 13/10/16
 # author: Camble
@@ -9,6 +5,15 @@ ptvsd.enable_attach('secret')
 # name: Safe-Power-Monitor - A utility to compliment the Safe Shutdown Switch for the Gameboy Zero project
 # description: A GPIO monitor with graceful shutdown capability and video warning overlays
 # source: https://github.com/Camble/Safe-Power-Monitor
+
+import RPi.GPIO as GPIO
+import subprocess
+import sys
+import time
+import os
+from datetime import timedelta
+from datetime import datetime
+from shutil import copyfile
 
 AdafruitPowerBoost  = True # Set this to False if using a generic power booster/charger module ie. a "BangGood" or "GearBest" Module
 DebugLog            = True # Set this to False to disable writing to the log file
@@ -30,21 +35,6 @@ lowalertVideo       = "~/Safe-Power-Monitor/lowbattalert.mp4"         # Alphanum
 
 # ==================== DO NOT CHANGE ANYTHING BELOW THIS LINE ====================
 
-import RPi.GPIO as GPIO
-import subprocess
-import sys
-import time
-import os
-from PowerWatcher import PowerWatcher
-from datetime import timedelta
-from datetime import datetime
-from shutil import copyfile
-
-if AdafruitPowerBoost is True:
-	from AdafruitBatteryWatcher import AdafruitBatteryWatcher
-else:
-	from BatteryWatcher import BatteryWatcher
-
 def log(code, message):
   if DebugLog is True:
     # If the log file doesn't exist, create it
@@ -54,6 +44,134 @@ def log(code, message):
     file = open(logFile, "a")
     file.write(datetime.now().strftime("%Y-%m-%d %H:%M:%S") + " [" + str(code) + "] " + message + "\n")
     file.close()
+
+class GpioWatcher(object):
+  def __init__(self, gpio_pin, internal_pull, trigger_state):
+    # Configure GPIO pin
+    self.pin = gpio_pin
+    self.trigger = trigger_state
+    if internal_pull is GPIO.PUD_DOWN:
+      self.pull = GPIO.PUD_DOWN
+    else:
+      self.pull = GPIO.PUD_UP
+
+    # Set edge type for event listener
+    if trigger_state is 0:
+      self.edge = GPIO.FALLING
+    elif trigger_state is 1:
+      self.edge = GPIO.RISING
+
+    # Create a threaded event listener
+    try:
+      GPIO.setup(self.pin, GPIO.IN, pull_up_down=self.pull)
+      GPIO.remove_event_detect(self.pin)
+      GPIO.add_event_detect(self.pin, self.edge, callback=self.callbackFunc, bouncetime=300)
+
+    except KeyboardInterrupt:
+      GPIO.cleanup()
+
+    # If the pin is already triggered, perform the callback
+    if GPIO.input(self.pin) is self.trigger:
+      self.callbackFunc()
+
+  def callbackFunc(self, channel):
+    log(99, "GPIO Pin " + str(self.pin) + " was triggered!")
+
+class PowerWatcher(GpioWatcher):
+  def callbackFunc(self, channel):
+    log(12, "Power switch on pin " + str(self.pin) + " initiated a shutdown.")
+    subprocess.call(['sudo shutdown -h now'], shell=True)
+    subprocess.call(['poweroff'], shell=True)
+    try:
+       sys.stdout.close()
+    except:
+       pass
+    try:
+       sys.stderr.close()
+    except:
+       pass
+    sys.exit(0)
+
+class BatteryWatcher(GpioWatcher):
+  def __init__(self, gpio_pin, internal_pull, trigger_state):
+    GpioWatcher.__init__(self, gpio_pin, internal_pull, trigger_state)
+    self.warnCount = 0
+    self.playerFlag = 0
+    self.previousWarn = None
+    self.callbackTriggered = 0
+
+  def warn(self):
+    # If the maximum warning count has been reached, skip it and shutdown
+    if self.warnCount >= numberOfWarnings:
+      shutdown()
+    else:
+      self.warnCount += 1
+      self.playerFlag = 1
+      self.previousWarn = time.time()
+      log(23, "Low battery warning number " + warnCount + " was displayed.")
+      os.system(videoPlayer + " " + lowalertVideo + " --alpha " + videoAlpha + ";")
+      playerFlag = 0
+
+      # Rebind GPIO event detect after system call (due to a bug with the GPIO library and threaded events)
+      GPIO.remove_event_detect(self.pin)
+      GPIO.add_event_detect(self.pin, self.edge, callback=self.callbackFunc, bouncetime=300)
+
+  def shutdown(self):
+    playerFlag = 1
+    os.system(videoPlayer + " " + shutdownVideo + " --alpha " + videoAlpha + ";");
+    playerFlag = 0
+    # Last chance to plug the charger in!
+    if GPIO.input(self.pin) is not trigger:
+      log(26, "Low battery on pin " + str(self.pin) + " was cancelled.")
+      return
+    else:
+      log(25, "Low battery on pin " + str(self.pin) + " initiated a shutdown.")
+      subprocess.call(['sudo shutdown -h now'], shell=True)
+      subprocess.call(['poweroff --poweroff'], shell=True)
+      sys.exit(0)
+
+  def monitor(self):
+    if self.callbackTriggered is 0 or self.playerFlag is 1:
+      return
+    if self.previousWarn is None:
+      self.warn()
+    else:
+      elapsed = time.time() - self.previousWarn
+      if elapsed >= 300:
+        self.warnCount = 0
+        self.playerFlag = 0
+        self.previousWarn = None
+        self.callbackTriggered = 0
+      elif elapsed >= 60:
+        self.warn()
+
+  def callbackFunc(self, channel):
+    if GPIO.input(self.pin) is not self.trigger:
+      self.callbackTriggered = 0
+      return
+
+    else:
+      self.callbackTriggered = 1
+      self.monitor()
+
+class BatteryWatcher_PB(BatteryWatcher):
+  def callbackFunc(self, channel):
+    # Checking for LED bounce for the duration of the battery timeout
+    for bounceSample in range(1, int(round(batteryTimeout / sampleRate))):
+      time.sleep(sampleRate)
+      if GPIO.input(self.pin) is not self.trigger:
+         break
+
+    while self.playerFlag is 1:
+      time.sleep(1)
+
+    # If the LED is a solid condition, there will be no bounce. Launch shutdown video and then gracefully shutdown
+    if bounceSample is int(round(batteryTimeout / sampleRate)) - 1:
+      self.shutdown()
+
+    # If the LED is a solid for more than 10% of the timeout, launch the low battery alert
+    if bounceSample > int(round(batteryTimeout / sampleRate * 0.1)):
+      self.monitor()
 
 def main():
   global logFile
@@ -135,15 +253,12 @@ def main():
 
   # Create some GpioWatchers
   powerWatcher = PowerWatcher(powerGPIO, powerInternalResistor, powerTriggerState)
-
   if (batteryGPIO is not None):
     playCount = 0
-    batteryWatcher = BatteryWatcher(batteryGPIO, batteryInternalResistor, batteryTriggerState)
-
-    #if (AdafruitPowerBoost is True):
-    #  batteryWatcher = AdafruitBatteryWatcher(batteryGPIO, batteryInternalResistor, batteryTriggerState)
-    #else:
-    #  batteryWatcher = BatteryWatcher(batteryGPIO, batteryInternalResistor, batteryTriggerState)
+    if (AdafruitPowerBoost is True):
+      batteryWatcher = BatteryWatcher_PB(batteryGPIO, batteryInternalResistor, batteryTriggerState)
+    else:
+      batteryWatcher = BatteryWatcher(batteryGPIO, batteryInternalResistor, batteryTriggerState)
 
 # Run the program
 main()
